fix: refactor for separation of concerns
